% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.comprehendmedical_package.R
\docType{package}
\name{paws.comprehendmedical-package}
\alias{paws.comprehendmedical}
\alias{paws.comprehendmedical-package}
\title{paws.comprehendmedical: AWS Comprehend Medical}
\description{
Comprehend Medical extracts structured information from
unstructured clinical text. Use these actions to gain insight in your
documents.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
